v2/requirements.txt
# Jarvis AI Agent v2.0 依赖包

# 核心依赖
selenium==4.15.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
undetected-chromedriver==3.5.4

# Web界面
streamlit==1.28.0

# AI服务
openai==1.3.0

# 数据处理
pandas==2.1.0

# 开发和测试
pytest==7.4.0
pytest-cov==4.1.0

# 日志和监控
colorlog==6.7.0

# 文件处理
pathlib2==2.3.7
